+++ crates/agentgateway/src/lib.rs
@@ -13,5 +13,4 @@ pub mod signal;
 pub mod sse;
 pub mod r#static;
 pub mod trcng;
-pub mod version;
 pub mod xds;
