+++ crates/core/src/lib.rs
@@ -2,5 +2,5 @@ pub mod drain;
 pub mod metrics;
 pub mod prelude;
 pub mod strng;
-// pub mod pool;
 pub mod copy;
+pub mod version;
